+++ Taehyung/AndroidStudy/.idea/misc.xml
@@ -19,7 +19,7 @@
         <entry key="..\:/Git/Android-Seminar-Study/Taehyung/AndroidStudy/features/Delivery/src/main/res/drawable/ic_my_baemin.xml" value="0.2574074074074074" />
         <entry key="..\:/Git/Android-Seminar-Study/Taehyung/AndroidStudy/features/Delivery/src/main/res/drawable/logo.xml" value="0.33125" />
         <entry key="..\:/Git/Android-Seminar-Study/Taehyung/AndroidStudy/features/Delivery/src/main/res/layout/activity_delivery_main.xml" value="0.19" />
-        <entry key="..\:/Git/Android-Seminar-Study/Taehyung/AndroidStudy/features/Delivery/src/main/res/layout/fragment_login.xml" value="0.33" />
+        <entry key="..\:/Git/Android-Seminar-Study/Taehyung/AndroidStudy/features/Delivery/src/main/res/layout/fragment_login.xml" value="0.25" />
         <entry key="..\:/Git/Android-Seminar-Study/Taehyung/AndroidStudy/features/Delivery/src/main/res/layout/fragment_main.xml" value="0.25" />
         <entry key="..\:/Git/Android-Seminar-Study/Taehyung/AndroidStudy/features/Delivery/src/main/res/layout/fragment_my.xml" value="0.4107692307692308" />
         <entry key="..\:/Git/Android-Seminar-Study/Taehyung/AndroidStudy/features/Delivery/src/main/res/layout/fragment_my_page.xml" value="0.25" />
