+++ Taehyung/AndroidStudy/features/Delivery/src/main/java/com/terry/delivery/base/BaseViewModel.kt
@@ -1,5 +1,7 @@
 package com.terry.delivery.base
 
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import io.reactivex.disposables.CompositeDisposable
 
@@ -9,6 +11,10 @@ import io.reactivex.disposables.CompositeDisposable
 abstract class BaseViewModel : ViewModel() {
 
     protected val disposable = CompositeDisposable()
+
+    protected val _isLoading = MutableLiveData<Boolean>()
+    val isLoading: LiveData<Boolean>
+        get() = _isLoading
 
     override fun onCleared() {
         disposable.dispose()
