+++ Taehyung/AndroidStudy/features/Delivery/src/main/java/com/terry/delivery/entity/RefreshToken.kt
@@ -1,4 +1,4 @@
-package com.terry.delivery.remote.entity
+package com.terry.delivery.entity
 
 data class RefreshToken(
     val access: String?,
