+++ Taehyung/AndroidStudy/features/Delivery/src/main/java/com/terry/delivery/entity/VerifyToken.kt
@@ -1,4 +1,4 @@
-package com.terry.delivery.remote.entity
+package com.terry.delivery.entity
 
 /*
  * Created by Taehyung Kim on 2021-09-08
