+++ Taehyung/AndroidStudy/features/Delivery/src/main/java/com/terry/delivery/ui/MainFragment.kt
@@ -1,8 +1,9 @@
-package com.terry.delivery
+package com.terry.delivery.ui
 
 import android.os.Bundle
 import android.view.View
 import androidx.navigation.fragment.findNavController
+import com.terry.delivery.R
 import com.terry.delivery.base.BaseFragment
 import com.terry.delivery.databinding.FragmentMainBinding
 import com.terry.delivery.extensions.view.setTopDrawable
