+++ Taehyung/AndroidStudy/features/Delivery/src/main/java/com/terry/delivery/ui/MyPageFragment.kt
@@ -1,9 +1,9 @@
-package com.terry.delivery
+package com.terry.delivery.ui
 
 import android.os.Bundle
 import android.view.View
 import androidx.navigation.fragment.findNavController
-import androidx.navigation.ui.setupWithNavController
+import com.terry.delivery.R
 import com.terry.delivery.base.BaseFragment
 import com.terry.delivery.databinding.FragmentMyPageBinding
 import com.terry.delivery.util.SnackbarUtil
