+++ Taehyung/AndroidStudy/features/Delivery/src/main/java/com/terry/delivery/ui/login/LoginFragment.kt
@@ -1,4 +1,4 @@
-package com.terry.delivery
+package com.terry.delivery.ui.login
 
 import android.os.Bundle
 import android.view.View
@@ -6,10 +6,10 @@ import androidx.core.content.ContextCompat
 import androidx.fragment.app.activityViewModels
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.ui.setupWithNavController
+import com.terry.delivery.R
 import com.terry.delivery.base.BaseFragment
 import com.terry.delivery.databinding.FragmentLoginBinding
 import com.terry.delivery.util.SnackbarUtil
-import com.terry.delivery.viewmodels.LoginViewModel
 
 /*
  * Created by Taehyung Kim on 2021-09-07
