+++ Taehyung/AndroidStudy/features/Delivery/src/main/res/layout/activity_delivery_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DeliveryMainActivity">
+    tools:context=".ui.DeliveryMainActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragment_container_view"
