+++ Taehyung/AndroidStudy/features/Delivery/src/main/res/navigation/nav_main.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/splashFragment"
-        android:name="com.terry.delivery.SplashFragment"
+        android:name="com.terry.delivery.ui.SplashFragment"
         android:label="SplashFragment"
         tools:layout="@layout/fragment_splash">
         <action
@@ -21,7 +21,7 @@
 
     <fragment
         android:id="@+id/mainFragment"
-        android:name="com.terry.delivery.MainFragment"
+        android:name="com.terry.delivery.ui.MainFragment"
         android:label="MainFragment"
         tools:layout="@layout/fragment_main">
 
@@ -57,25 +57,25 @@
 
     <fragment
         android:id="@+id/searchFragment"
-        android:name="com.terry.delivery.SearchFragment"
+        android:name="com.terry.delivery.ui.SearchFragment"
         android:label="SearchFragment"
         tools:layout="@layout/fragment_search" />
 
     <fragment
         android:id="@+id/favoriteFragment"
-        android:name="com.terry.delivery.FavoriteFragment"
+        android:name="com.terry.delivery.ui.FavoriteFragment"
         android:label="FavoriteFragment"
         tools:layout="@layout/fragment_favorite" />
 
     <fragment
         android:id="@+id/orderListFragment"
-        android:name="com.terry.delivery.OrderListFragment"
+        android:name="com.terry.delivery.ui.OrderListFragment"
         android:label="OrderListFragment"
         tools:layout="@layout/fragment_order_list" />
 
     <fragment
         android:id="@+id/myPageFragment"
-        android:name="com.terry.delivery.MyPageFragment"
+        android:name="com.terry.delivery.ui.MyPageFragment"
         android:label="@string/my_page"
         tools:layout="@layout/fragment_my_page">
 
@@ -90,7 +90,7 @@
 
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.terry.delivery.LoginFragment"
+        android:name="com.terry.delivery.ui.login.LoginFragment"
         android:label="LoginFragment"
         tools:layout="@layout/fragment_login" />
